requirements.txt
Authlib==1.3.0
board==1.0
Flask==3.0.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.1.1
PyJWT==2.8.0
Requests==2.31.0
SQLAlchemy==2.0.25
